Rename storage related classes
