+++ bin/update-pdp
@@ -16,20 +16,14 @@ declare(strict_types=1);
 
 use Pdp\Storage\Logger;
 use Pdp\Storage\Psr16FileCache;
-use Pdp\Storage\RulesCachePsr16Adapter;
-use Pdp\Storage\TopLevelDomainsCachePsr16Adapter;
+use Pdp\Storage\PublicSuffixListCachePsr16Adapter;
+use Pdp\Storage\RootZoneDatabaseCachePsr16Adapter;
 use Pdp\Storage\CurlHttpClient;
 use Pdp\Storage\Manager;
 
 const KEY_CACHE_DIR = 'cache-dir';
-const KEY_URI_PSL = 'psl';
-const KEY_URI_RZD = 'rzd';
 const KEY_TTL = 'ttl';
-const DEFAULT_VALUES = [
-    KEY_URI_PSL => 'https://publicsuffix.org/list/public_suffix_list.dat',
-    KEY_URI_RZD => 'https://data.iana.org/TLD/tlds-alpha-by-domain.txt',
-    KEY_TTL => '1 DAY',
-];
+const DEFAULT_TTL = '1 DAY';
 
 function success(string $message): void
 {
@@ -80,8 +74,6 @@ $arguments = getopt('h::', [
     'h',
     'help',
     KEY_CACHE_DIR.':',
-    KEY_URI_PSL.':',
-    KEY_URI_RZD.':',
     KEY_TTL.':',
 ]);
 
@@ -92,8 +84,6 @@ if (isset($arguments['help']) || isset($arguments['h']) || [] === $arguments) {
     $script [options]
 
 {$yellow}Options:$reset
-$green    --%s=URL$reset        set the URL to use to refresh the Public Suffix List ({$yellow}default:$reset '%s')
-$green    --%s=URL$reset        set the URL to use to refresh the Root Zone Database ({$yellow}default:$reset '%s')
 $green    --%s=TTL$reset        set the TTL for the cache ({$yellow}default:$reset '%s')
 $green    --%s=PATH$reset set the cache root directory ({$yellow}default:$reset '$default_cache_dir')
 $green    -h, --help$reset       show the following help message
@@ -108,20 +98,13 @@ $green    -h, --help$reset       show the following help message
    Refresh all the cache using another cache directory
    $green$script --cache-dir=/temp$reset
 
-   Refresh all the cache using another Public Suffix List URI but with the default URI for the Root Zone Database
-   $green$script --psl="https://example.com/public-suffix-list"$reset
-
    Read more at https://github.com/jeremykendall/php-domain-parser/
 HELP;
 
     success(sprintf(
         $helpText,
-        KEY_URI_PSL,
-        DEFAULT_VALUES[KEY_URI_PSL],
-        KEY_URI_RZD,
-        DEFAULT_VALUES[KEY_URI_RZD],
         KEY_TTL,
-        DEFAULT_VALUES[KEY_TTL],
+        DEFAULT_TTL,
         KEY_CACHE_DIR
     ));
 
@@ -134,9 +117,7 @@ if (!extension_loaded('curl')) {
     die(1);
 }
 
-$context = filter_var_array($arguments + [KEY_CACHE_DIR => $default_cache_dir] + DEFAULT_VALUES, [
-    KEY_URI_PSL => ['filter' => FILTER_SANITIZE_STRING, 'flags' => FILTER_FLAG_STRIP_LOW, 'default' => ''],
-    KEY_URI_RZD => ['filter' => FILTER_SANITIZE_STRING, 'flags' => FILTER_FLAG_STRIP_LOW, 'default' => ''],
+$context = filter_var_array($arguments + [KEY_CACHE_DIR => $default_cache_dir, KEY_TTL => DEFAULT_TTL], [
     KEY_TTL => ['filter' => FILTER_SANITIZE_STRING, 'flags' => FILTER_FLAG_STRIP_LOW],
     KEY_CACHE_DIR => ['filter' => FILTER_SANITIZE_STRING, 'flags' => FILTER_FLAG_STRIP_LOW],
 ]);
@@ -146,8 +127,8 @@ try {
     $cache = new Psr16FileCache($context[KEY_CACHE_DIR]);
     $manager = new Manager(
         new CurlHttpClient(),
-        new RulesCachePsr16Adapter($cache, $context[KEY_TTL], $logger),
-        new TopLevelDomainsCachePsr16Adapter($cache, $context[KEY_TTL], $logger),
+        new PublicSuffixListCachePsr16Adapter($cache, $context[KEY_TTL], $logger),
+        new RootZoneDatabaseCachePsr16Adapter($cache, $context[KEY_TTL], $logger),
     );
 
 } catch (Throwable $exception) {
@@ -161,26 +142,21 @@ try {
 
 success("$yellow Starting updating your local cache.$reset");
 $error = 0;
-$pslUri = trim($context[KEY_URI_PSL]);
-if ('' !== $pslUri) {
-    try {
-        $manager->getPublicSuffixListRemoteCopy($pslUri);
-    } catch (Throwable $exception) {
-        $logger->error('Local cache update failed with {exception}', ['exception' => $exception->getMessage()]);
-        fail("$redbg The command failed to update the public suffix list local cache. $reset");
-        ++$error;
-    }
+
+try {
+    $manager->getPublicSuffixListRemoteCopy();
+} catch (Throwable $exception) {
+    $logger->error('Local cache update failed with {exception}', ['exception' => $exception->getMessage()]);
+    fail("$redbg The command failed to update the public suffix list local cache. $reset");
+    ++$error;
 }
 
-$rzdUri = trim($context[KEY_URI_RZD]);
-if ('' !== $rzdUri) {
-    try {
-        $manager->getRootZoneDatabaseRemoteCopy($context[KEY_URI_RZD]);
-    } catch (Throwable $exception) {
-        $logger->error('Local cache update failed with {exception}', ['exception' => $exception->getMessage()]);
-        fail("$redbg The command failed to update the root zone database local cache. $reset");
-        ++$error;
-    }
+try {
+    $manager->getRootZoneDatabaseRemoteCopy();
+} catch (Throwable $exception) {
+    $logger->error('Local cache update failed with {exception}', ['exception' => $exception->getMessage()]);
+    fail("$redbg The command failed to update the root zone database local cache. $reset");
+    ++$error;
 }
 
 if (0 === $error) {
