+++ src/Storage/PublicSuffixListCache.php
@@ -15,30 +15,21 @@ declare(strict_types=1);
 
 namespace Pdp\Storage;
 
-use Pdp\RootZoneDatabaseInterface;
+use Pdp\Rules;
 
-/**
- * Top Level Domains Cache Provider.
- *
- * This class obtains, writes, caches, and returns PHP representations
- * of the Public Suffix List ICANN section
- *
- * @author Jeremy Kendall <jeremy@jeremykendall.net>
- * @author Ignace Nyamagana Butera <nyamsprod@gmail.com>
- */
-interface TopLevelDomainsCache
+interface PublicSuffixListCache
 {
     /**
-     * Gets the Top Level Domains from the Cache.
+     * Gets the Public Suffix List Rules.
      */
-    public function fetchByUri(string $uri): ?RootZoneDatabaseInterface;
+    public function fetchByUri(string $uri): ?Rules;
 
     /**
-     * Cache the Top Level Domains List.
+     * Cache the Public Suffix List Rules.
      *
      * If a local cache already exists, it will be overwritten.
      *
      * Returns true if the refresh was successful, false otherwise
      */
-    public function storeByUri(string $uri, RootZoneDatabaseInterface $topLevelDomains): bool;
+    public function storeByUri(string $uri, Rules $rules): bool;
 }
